+++ src/main/java/exemplojdbc/JdbcExemplo3LocalizarListarTabelaGenerica.java
@@ -68,7 +68,7 @@ public class JdbcExemplo3LocalizarListarTabelaGenerica extends Base {
             statement.setString(1, uf);
             final var result = statement.executeQuery();
             if(result.next()){
-                System.out.printf("Id: %2d Nome: %-30s UF: %s\n", result.getInt("id"), result.getString("nome"), result.getString("uf"));
+                System.out.printf("Id: %2d Nome: %-30s UF: %s\n", result.getLong("id"), result.getString("nome"), result.getString("uf"));
             }
             System.out.println();
         } catch(SQLException e){
